test: separate test class
