+++ src/main/java/racingcar/domain/car/Car.java
@@ -1,4 +1,4 @@
-package racingcar.domain.Car;
+package racingcar.domain.car;
 
 import java.util.Objects;
